refactor: separate file for makeUrlAbsolute
