+++ packages/api-reference/src/helpers/index.ts
@@ -12,6 +12,7 @@ export * from './hasModels'
 export * from './hasSecuritySchemes'
 export * from './hasWebhooks'
 export * from './joinWithSlash'
+export * from './makeUrlAbsolute'
 export * from './mapFromObject'
 export * from './mergeAllObjects'
 export * from './parse'
